compiles clean with -Wall
Ignore-this: e9a6e7c9fc6e58d2c884bddd8e8bfa75

Now everything compiles clean with -Wall -Werror.
Compiler found a couple of missed return statements. nice!

darcs-hash:20141001033919-c21fe-9cd355680661b01413ad4f8df816c7c4ba8615df

+++ src/network/connection_manager.hpp
@@ -43,7 +43,7 @@ namespace fire
 {
     namespace network 
     {
-        using assignment_map = std::unordered_map<std::string, int>; 
+        using assignment_map = std::unordered_map<std::string, size_t>; 
         using tcp_connection_pool = std::vector<tcp_queue_ptr>;
         using connection_map = std::unordered_map<std::string, connection*>;
